api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cupy-cuda12x>=13.3.0",
    "fastapi>=0.115.2",
    "gunicorn>=23.0.0",
    "pillow>=10.4.0",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.31.1",
]
